+++ src/ledgerapp.spec.ts
@@ -1,5 +1,5 @@
 import { Secp256k1, Secp256k1Signature, Sha256 } from "@iov/crypto";
-import { Encoding, isUint8Array } from "@iov/encoding";
+import { Encoding } from "@iov/encoding";
 import Transport from "@ledgerhq/hw-transport";
 import TransportNodeHid from "@ledgerhq/hw-transport-node-hid";
 import * as semver from "semver";
@@ -13,7 +13,6 @@ import {
 import { IovLedgerApp, isIovLedgerAppAddress, isIovLedgerAppSignature, isIovLedgerAppVersion } from "./ledgerapp";
 
 const { fromHex } = Encoding;
-const { fromDer } = Secp256k1Signature;
 
 describe("IovLedgerApp", () => {
   // https://github.com/Zondax/ledger-cosmos-js/blob/master/tests/basic.ispec.js#L132
@@ -73,6 +72,21 @@ describe("IovLedgerApp", () => {
   });
 
   describe("getAddress", () => {
+    it("can get address with custom HRP", async () => {
+      pendingWithoutSeededLedger();
+
+      const app = new IovLedgerApp(transport!, "cosmos");
+      const version = await app.getVersion();
+      if (!isIovLedgerAppVersion(version)) throw new Error(version.errorMessage);
+
+      const response = await app.getAddress(0);
+      if (!isIovLedgerAppAddress(response)) throw new Error(response.errorMessage);
+
+      // Dave's Nano S
+      expect(response.pubkey).toEqual(fromHex("03910af2b917aa3f3d47cbd5acd00515c4254942bd6d288f6c143c58b59c754125"));
+      expect(response.address).toEqual("cosmos1y4t33z7ugz2323vnuhjwftz33quns0t4rfdhku");
+    });
+
     it("can get address", async () => {
       pendingWithoutSeededLedger();
 
@@ -139,7 +153,7 @@ describe("IovLedgerApp", () => {
   });
 
   describe("sign", () => {
-    it("can sign", async () => {
+    it("can sign string", async () => {
       pendingWithoutSeededLedger();
       pendingWithoutInteractiveLedger();
 
@@ -154,12 +168,61 @@ describe("IovLedgerApp", () => {
       const responseSign = await app.sign(accountIndex, message);
       if (!isIovLedgerAppSignature(responseSign)) throw new Error(responseSign.errorMessage);
 
-      expect(isUint8Array(responseSign.signature)).toEqual(true);
+      expect(responseSign.signature instanceof Secp256k1Signature).toEqual(true);
 
       // Check signature is valid
       const encoded = Uint8Array.from(message.split(""), s => s.charCodeAt(0));
       const prehash = new Sha256(encoded).digest();
-      const valid = await Secp256k1.verifySignature(fromDer(responseSign.signature), prehash, responseAddr.pubkey);
+      const valid = await Secp256k1.verifySignature(responseSign.signature, prehash, responseAddr.pubkey);
+
+      expect(valid).toEqual(true);
+    });
+
+    it("can sign object", async () => {
+      pendingWithoutSeededLedger();
+      pendingWithoutInteractiveLedger();
+
+      const app = new IovLedgerApp(transport!);
+      const version = await app.getVersion();
+      if (!isIovLedgerAppVersion(version)) throw new Error(version.errorMessage);
+
+      const accountIndex = 0;
+      const responseAddr = await app.getAddress(accountIndex);
+      if (!isIovLedgerAppAddress(responseAddr)) throw new Error(responseAddr.errorMessage);
+      const tx = {
+        // NOTE: Ledger app expects `msgs` but the chain expects `msg`
+        msgs: [
+          {
+            type: "domain/RegisterDomain",
+            value: {
+              domain: "registered_via_ledger",
+              admin: "star1478t4fltj689nqu83vsmhz27quk7uggjwe96yk",
+              type: "closed",
+              broker: "",
+              // eslint-disable-next-line @typescript-eslint/camelcase
+              fee_payer: "",
+            },
+          },
+        ],
+        fee: { amount: [{ denom: "uvoi", amount: "2000000" }], gas: "200000" },
+        signatures: null,
+        memo: "this is the memo",
+        // eslint-disable-next-line @typescript-eslint/camelcase
+        account_number: "123",
+        // eslint-disable-next-line @typescript-eslint/camelcase
+        chain_id: "iovns-galaxynet",
+        sequence: "456",
+      };
+      const responseSign = await app.sign(accountIndex, tx);
+      if (!isIovLedgerAppSignature(responseSign)) throw new Error(responseSign.errorMessage);
+
+      expect(responseSign.signature instanceof Secp256k1Signature).toEqual(true);
+
+      // Check signature is valid
+      const sorted = JSON.stringify(IovLedgerApp.sortObject(tx));
+      const encoded = Uint8Array.from(sorted.split(""), s => s.charCodeAt(0));
+      const prehash = new Sha256(encoded).digest();
+      const valid = await Secp256k1.verifySignature(responseSign.signature, prehash, responseAddr.pubkey);
 
       expect(valid).toEqual(true);
     });